[#2] refactor: APIParams 인터페이스의 위치를 type.ts으로 이동
